+++ pkg/test/statefulsets.go
@@ -55,7 +55,7 @@ func StatefulSet(overrides ...StatefulSetOptions) *appsv1.StatefulSet {
 	return &appsv1.StatefulSet{
 		ObjectMeta: objectMeta,
 		Spec: appsv1.StatefulSetSpec{
-			Replicas: lo.ToPtr(int32(options.Replicas)),
+			Replicas: lo.ToPtr(options.Replicas),
 			Selector: &metav1.LabelSelector{MatchLabels: options.PodOptions.Labels},
 			Template: v1.PodTemplateSpec{
 				ObjectMeta: ObjectMeta(options.PodOptions.ObjectMeta),
